+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Features
 - Added default run name handling ([#18](https://github.com/neptune-ai/neptune-client-experimental/pull/18))
 
+### Changes
+- Add custom units and descriptions to tqdm progress bars ([#29](https://github.com/neptune-ai/neptune-client-experimental/pull/29))
+
 
 ## neptune-experimental 0.2.1